Task Completed

+++ src/Pages/Quiz.jsx
@@ -1,30 +1,30 @@
 import React, { useEffect, useState } from "react";
 import Styles from "../assets/css/quiz.module.css";
 import Timer from "../components/Timer/Timer";
+import NextBtn from "../components/Button/Next";
+import AlertBox from "../components/AlertBox";
 import QuiestionTimer from "../components/Timer/QuestionTimer";
 import { collection, getDocs, query } from "firebase/firestore";
 import { db } from "../firebase";
-
 import { Form } from "react-bootstrap";
-import NextBtn from "../components/Button/Next";
 import { SyncLoader } from "react-spinners";
-import FinishBtn from "../components/Button/FinishBtn";
 import { useNavigate } from "react-router-dom";
 
 const Quiz = ({ loading, setLoading, marks, setMarks }) => {
-  const [time, setTime] = useState(60);
   const [data, setData] = useState([]);
+  const [currentQuestionIndex, setCurrentQuestionIndex] = useState(0);
+  const [currentData, setCurrentData] = useState();
   const [questionNo, setQuestionNo] = useState(null);
   const [selectData, setSelectdata] = useState([]);
+  // eslint-disable-next-line
   const [answers, setAnswers] = useState([]);
-
-  console.log("answer", answers);
+  const [showAlert, setShowAlert] = useState(false);
 
   const navigate = useNavigate();
 
   var userName = localStorage.getItem("User");
 
-  const q = query(collection(db, "Questions"));
+  const q = query(collection(db, "Current"));
 
   const getData = async () => {
     setLoading(true);
@@ -34,10 +34,9 @@ const Quiz = ({ loading, setLoading, marks, setMarks }) => {
     querySnapshot.forEach((doc) => {
       newData.push(doc.data());
     });
-    setData(newData[(Math.random() * newData.length) | 0]);
+    setData(newData);
     setQuestionNo(questionNo + 1);
     setLoading(false);
-    setTime(60);
   };
 
   useEffect(() => {
@@ -45,60 +44,79 @@ const Quiz = ({ loading, setLoading, marks, setMarks }) => {
     // eslint-disable-next-line
   }, []);
 
-  const clickHandler = () => {
+  useEffect(() => {
+    setCurrentData(data[currentQuestionIndex]);
+    // eslint-disable-next-line
+  }, [data.length, currentQuestionIndex]);
+
+  const clickHandler = async () => {
     setAnswers((prev) => [...prev, selectData]);
     getData();
     setData([]);
+    setCurrentQuestionIndex(currentQuestionIndex + 1);
     setSelectdata([]);
-    if (selectData.ans === data.Answer) {
+    if (selectData.ans === currentData.Answer) {
       setMarks(marks + 1);
     }
-  };
-
-  useEffect(() => {
-    if (time === 0) {
-      clickHandler();
+    if (questionNo === data.length) {
+      navigate("/end");
+      setTimeout(function () {
+        localStorage.clear();
+      }, 1000);
     }
-    // eslint-disable-next-line
-  }, [time]);
-
-  const finishTask = () => {
-    navigate("/end");
-    localStorage.clear();
   };
 
   const optionChange = (id, e) => {
     setSelectdata({ ans: e.target.value, id });
   };
 
+  const forceEnd = () => {
+    navigate("/end");
+    setTimeout(function () {
+      localStorage.clear();
+    }, 500);
+  };
+
+  document.body.addEventListener("mouseleave", (e) => {
+    setShowAlert(true);
+  });
+
+  window.onbeforeunload = function (e) {
+    localStorage.clear();
+  };
+
   return (
     <div>
+      {showAlert && (
+        <AlertBox forceEnd={forceEnd} setShowAlert={setShowAlert} />
+      )}
       <div className={Styles.heading}>
         <h3> Welcome {userName}</h3>
-
         <h2>Quiz</h2>
         <div className={Styles.timer}>
           <Timer />
         </div>
       </div>
       <div className={Styles.quiz_container}>
-        {data.id && data.id ? (
+        {currentData && currentData.id ? (
           <div>
             <div className={Styles.head}>
-              <h4> Question {questionNo} of 10</h4>
+              <h4>
+                Question {questionNo} of {data.length}
+              </h4>
             </div>
             <div className={Styles.timer}>
-              <QuiestionTimer time={time} setTime={setTime} />
+              <QuiestionTimer clickHandler={clickHandler} />
             </div>
             <div className={Styles.display_data}>
               <h3>
                 {questionNo && `${questionNo} : `}
-                {data.Question}
+                {currentData.Question}
               </h3>
               <br />
               <Form className={Styles.option}>
-                {data.Option &&
-                  data.Option.map((option, i) => {
+                {currentData.Option &&
+                  currentData.Option.map((option, i) => {
                     return (
                       <div key={i} className="mb-3">
                         <Form.Group controlId="kindOfStand">
@@ -108,8 +126,7 @@ const Quiz = ({ loading, setLoading, marks, setMarks }) => {
                             value={option}
                             name="group1"
                             type="radio"
-                            onChange={(e) => optionChange(data.id, e)}
-                            // checked={option === answer.ans}
+                            onChange={(e) => optionChange(currentData.id, e)}
                           />
                         </Form.Group>
                       </div>
@@ -117,20 +134,12 @@ const Quiz = ({ loading, setLoading, marks, setMarks }) => {
                   })}
               </Form>
 
-              {questionNo === 10 ? (
-                <FinishBtn
-                  name="Finish"
-                  finishTask={finishTask}
-                  loading={loading}
-                />
-              ) : (
-                <NextBtn
-                  name="Next"
-                  loading={loading}
-                  clickHandler={clickHandler}
-                  disabled={!selectData.ans ? true : false}
-                />
-              )}
+              <NextBtn
+                name={questionNo === data.length ? "Finish" : "Next"}
+                loading={loading}
+                clickHandler={clickHandler}
+                disabled={!selectData.ans ? true : false}
+              />
             </div>
           </div>
         ) : (
